move tap types to voiceTap
